// src/core/fhir/types/fhirObservation.ts
export interface FhirObservation extends FhirResource {
// ...
  status: 'registered' | 'preliminary' | 'final' | 'amended' | 'corrected' | 'cancelled' | 'entered-in-error' | 'unknown';
// ...
}
// ...
export interface VitalSignsObservation extends FhirObservation {
  category: [{
    coding: [{
      system: 'http://terminology.hl7.org/CodeSystem/observation-category';
      code: 'vital-signs';
      display: 'Vital Signs';
    }];
    text: 'Vital Signs';
  }];
}